Kill some compiler warnings under Linux and mingw32
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS@4994 212acab6-be3b-0410-9dea-997c60f758d6

+++ trunk/Contrib/Makensisw/toolbar.cpp
@@ -216,7 +216,7 @@ void ShowToolbarDropdownMenu()
   RECT rect;
   GetWindowRect(g_toolbar.hwnd, (LPRECT) &rect);
   TrackPopupMenu(g_toolbar.dropdownmenu,
-                 NULL,
+                 (UINT)NULL,
                  rect.left + (int)(short)g_toolbar.dropdownpoint.x,
                  rect.top + (int)(short)g_toolbar.dropdownpoint.y,
                  0,
